f nightly
argh.

The errors are all still with middleware (at least the ones that rust
feels like telling me about right now).

I really think that actix does NOT want you doing anything in middleware
anymore.

in server/mod.rs:
    the `wrap` function is demanding that there's an unsatisfied trait
`IntoTransform` missing. The only existing middleware that uses it is
`cors.rs` and that has more factories in it than Java. I was basing
things off of `defaultheaders.rs` which has a Transform, but no
IntoTransform.

The alternate method is to use `wrap_fn`, and pass it to a fn that does
the same operations as the `*Middleware.call()` fn, but that gets into a
huge rabbit hole of type definitions that are gleefully skipped over by
the anonymous fn wrapper.

Inside of middleware, it's sort of the same thing, where there are
missing Future traits on boxed ServiceResponses

(ノ°Д°）ノ︵ ┻━┻