Partially fix batch handling.
There are a few type errors that need to be fixed, but I think the general approach is correct.

This assumes that the batches table has a new "timestamp" field, which I added to the test database. Since id is used as part of the primary key, it can't contain duplicates. The code generates a unique timestamp for every bso in the batch to use as an id, but then selects based on the timestamp field instead of the id field.